Made some modifications
