+++ modules/display_time.js
@@ -1,0 +1,7 @@
+export const displayTime = () => {
+    const time = document.getElementById('date');
+    const currentDate = luxon.DateTime.local().toLocaleString(
+        luxon.DateTime.DATETIME_FULL,
+    );
+    time.innerHTML = currentDate;
+}
